updated the provenance for file
put it into a function that we can apply whenever we open a nc file
